requirements.txt
pdfplumber
pdf2image
requests
pinecone-client
python-dotenv
fastapi
pydantic
uvicorn
